# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "EvidenceCandidateRequest": {
        "properties": {
          "candidate_name": {
            "minLength": 1,
            "title": "Candidate Name",
            "type": "string"
          },
          "cost": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Cost",
            "type": "number"
          },
          "decision_value": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Decision Value",
            "type": "number"
          },
          "evidence_id": {
            "minLength": 1,
            "title": "Evidence Id",
            "type": "string"
          },
          "feasibility": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Feasibility",
            "type": "number"
          },
          "infeasibility_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Infeasibility Reason"
          },
          "is_feasible": {
            "default": true,
            "title": "Is Feasible",
            "type": "boolean"
          },
          "reliability": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Reliability",
            "type": "number"
          },
          "time": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Time",
            "type": "number"
          }
        },
        "required": [
          "evidence_id",
          "candidate_name",
          "decision_value",
          "reliability",
          "feasibility",
          "cost",
          "time"
        ],
        "title": "EvidenceCandidateRequest",
        "type": "object"
      },
...
      "EvidenceGraphRecordRequest": {
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "decision_id": {
            "minLength": 1,
            "title": "Decision Id",
            "type": "string"
          },
          "graph_id": {
            "minLength": 1,
            "title": "Graph Id",
            "type": "string"
          },
          "incident_id": {
            "minLength": 1,
            "title": "Incident Id",
            "type": "string"
          },
          "rationale": {
            "minLength": 1,
            "title": "Rationale",
            "type": "string"
          },
          "relationship": {
            "minLength": 1,
            "title": "Relationship",
            "type": "string"
          },
          "source_node": {
            "minLength": 1,
            "title": "Source Node",
            "type": "string"
          },
          "source_type": {
            "minLength": 1,
            "title": "Source Type",
            "type": "string"
          },
          "state": {
            "minLength": 1,
            "title": "State",
            "type": "string"
          },
          "target_node": {
            "minLength": 1,
            "title": "Target Node",
            "type": "string"
          },
          "target_type": {
            "minLength": 1,
            "title": "Target Type",
            "type": "string"
          }
        },
        "required": [
          "graph_id",
          "incident_id",
          "decision_id",
          "source_node",
          "source_type",
          "relationship",
          "target_node",
          "target_type",
          "confidence",
          "state",
          "rationale"
        ],
        "title": "EvidenceGraphRecordRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IntelligenceOverviewRequest": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/EvidenceCandidateRequest"
            },
            "minItems": 1,
            "title": "Candidates",
            "type": "array"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/EvidenceGraphRecordRequest"
            },
            "minItems": 1,
            "title": "Records",
            "type": "array"
          }
        },
        "required": [
          "records",
          "candidates"
        ],
        "title": "IntelligenceOverviewRequest",
        "type": "object"
      },
      "IntelligenceOverviewResponse": {
        "properties": {
          "actors": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Actors",
            "type": "array"
          },
          "gaps": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Gaps",
            "type": "array"
          },
          "graph": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Graph",
            "type": "array"
          },
          "hypotheses": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Hypotheses",
            "type": "array"
          },
          "profiles": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Profiles",
            "type": "array"
          },
          "ranking": {
            "$ref": "#/components/schemas/RankEvidenceResponse"
          },
          "states": {
            "items": {
              "$ref": "#/components/schemas/IntelligenceStateResponse"
            },
            "title": "States",
            "type": "array"
          }
        },
        "required": [
          "graph",
          "states",
          "gaps",
          "ranking"
        ],
        "title": "IntelligenceOverviewResponse",
        "type": "object"
      },
      "IntelligenceStateResponse": {
        "properties": {
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "evidence_id",
          "state",
          "reason"
        ],
        "title": "IntelligenceStateResponse",
        "type": "object"
      },
...
      "RankEvidenceResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/RankedEvidenceResponse"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results"
        ],
        "title": "RankEvidenceResponse",
        "type": "object"
      },
      "RankedEvidenceResponse": {
        "properties": {
          "candidate_name": {
            "title": "Candidate Name",
            "type": "string"
          },
          "cost": {
            "title": "Cost",
            "type": "number"
          },
          "decision_value": {
            "title": "Decision Value",
            "type": "number"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "feasibility": {
            "title": "Feasibility",
            "type": "number"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "reliability": {
            "title": "Reliability",
            "type": "number"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "selected": {
            "title": "Selected",
            "type": "boolean"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "title": "Strengths",
            "type": "array"
          },
          "time": {
            "title": "Time",
            "type": "number"
          },
          "tradeoffs": {
            "items": {
              "type": "string"
            },
            "title": "Tradeoffs",
            "type": "array"
          }
        },
        "required": [
          "evidence_id",
          "candidate_name",
          "score",
          "rank",
          "decision_value",
          "reliability",
          "feasibility",
          "cost",
          "time",
          "explanation",
          "strengths",
          "tradeoffs",
          "selected"
        ],
        "title": "RankedEvidenceResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Decision-Aware Evidence Orchestration for One Health",
    "title": "AquaPass API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/intelligence/incidents/{incident_id}/overview": {
      "get": {
        "description": "Compose current database evidence with the deterministic intelligence modules.",
        "operationId": "get_incident_intelligence_overview_api_intelligence_incidents__incident_id__overview_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntelligenceOverviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incident Intelligence Overview",
        "tags": [
          "intelligence"
        ]
      }
    },
    "/api/intelligence/overview": {
      "post": {
        "description": "Return graph, state, gap and ranking data in one frontend call.\n\nSang's deterministic graph and ranking modules remain the source of truth;\nthis route only composes their existing contracts for the decision workspace.",
        "operationId": "intelligence_overview_api_intelligence_overview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntelligenceOverviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntelligenceOverviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Intelligence Overview",
        "tags": [
          "intelligence"
        ]
      }
    },
...
  }
}